Adding Range Support to Salt Range is a cluster based metadata store.  It was used internally at yahoo and recently open source.  You can read about range here:
https://github.com/grierj/range/wiki/Introduction-to-Range-with-YAML-files

Range allows arbitrary grouping of hosts via clusters and the storage
of metadata along with them.  Additionally there is a powerful DSL that
allows for set operations, regexes, and multilayer cluster referencing.

The implementation here runs the range query and reduces it to a list
style expression format, which means only the salt master needs to know
about range or have the range libraries installed.