Back-port #33613 to 2016.3 (#33638)
* Updated apache_module for backward compatible.

[This](https://github.com/saltstack/salt/commit/eac6d1a7f16bedd2f66d8612ec7bf32ac91f161a) introduced
a change of functions names that break this module when using Beryllium or older and upgrading to Boron.
Added here a deprecated warning in to order to advice users to migrate to new functions names. Also
documentation updated since was not updated in the original commit.

* Move warnutil from Carbon to Nitrogen

Let's give people a little more time to update their states.